app/requirements.txt
faiss-cpu==1.7.4
pandas==2.2.2
sentence-transformers==2.5.1
torch==2.1.2
transformers==4.40.0
sentencepiece==0.2.0
numpy==1.26.4



fastapi
pydantic
uvicorn